backend/pyproject.toml
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "currency_exchanger.settings"
addopts = "--reuse-db"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | \.hg
  | \.venv
  | \.mypy_cache
  | _build
  | build
  | dist
  | env
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["typeland", "tests"]
skip = ["migrations"]

[tool.poetry]
name = "currency-exchanger"
version = "0.1.0"
description = "Web application for currency exchanges, money transfers and stock trading."
authors = [
    "Norbert Cyran <cyran.norbert97@gmail.com>",
    "Maciej Jaskuła",
    "Dominik Babiński",
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.2"
djangorestframework = "^3.12.1"
dj-rest-auth = {extras = ["with_social"], version = "^1.1.2"}
django-cors-headers = "^3.5.0"
psycopg2 = "^2.8.6"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
pytest-django = "^3.10.0"
pytest-mock = "^3.3.1"
black = "^20.8b1"
isort = "^5.6.3"
pre-commit = "^2.7.1"
django-extensions = "^3.0.9"
model-bakery = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
